# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce AI Support API",
    "version": "1.0.0",
    "description": "REST API for order management with AI-powered support agent"
  },
  "paths": {
    "/api/orders": {
      "post": {
        "summary": "Create an order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["items"],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["productId", "quantity", "unitPrice"],
                      "properties": {
                        "productId": { "type": "string" },
                        "quantity": { "type": "integer" },
                        "unitPrice": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created" },
          "400": { "description": "Invalid input" }
        }
      },
      "get": {
        "summary": "List all orders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["PENDING", "PROCESSING", "SHIPPED", "DELIVERED"] }
          }
        ],
        "responses": {
          "200": { "description": "List of orders" }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Get order details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Order details" },
          "404": { "description": "Order not found" }
        }
      },
      "delete": {
        "summary": "Cancel an order",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Order cancelled" },
          "404": { "description": "Order not found" },
          "409": { "description": "Order cannot be cancelled (not PENDING)" }
        }
      }
    },
    "/api/ai/chat": {
      "post": {
        "summary": "AI Support Agent",
        "description": "Natural language chat endpoint. The AI processes user input, cross-references with order status and corporate policies, and can autonomously cancel orders via tool calling.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": { "type": "string" },
                  "orderId": { "type": "string", "description": "Optional order ID for context" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "AI response" },
          "500": { "description": "Internal error" }
        }
      }
    },
    "/api/ai/logs": {
      "get": {
        "summary": "AI Observability Logs",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "intent", "in": "query", "schema": { "type": "string" } },
          { "name": "injection", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "AI logs" }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Health Check",
        "responses": {
          "200": { "description": "Service healthy" },
          "503": { "description": "Service degraded" }
        }
      }
    }
  }
}